LayoutTests/imported/w3c/web-platform-tests/trusted-types/require-trusted-types-for-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.


FAIL Require trusted types for 'script' block create HTML. assert_throws_js: function "_ => {
          d.innerHTML = "a";
        }" did not throw
FAIL Require trusted types for 'script' block create script. assert_throws_js: function "_ => {
          d.innerText = "a";
        }" did not throw
FAIL Require trusted types for 'script' block create script URL. assert_throws_js: function "_ => {
          s.src = "a";
        }" did not throw
PASS Set require trusted types for 'script' without CSP for trusted types don't block policy creation and using.
